+++ src/printer_ast.ml
@@ -184,6 +184,9 @@ let to_const = function
   | Cmax            -> "max"
   | Cmin            -> "min"
   | Csubsetof       -> "subsetof"
+  | Cconcat         -> "concat"
+  | Cslice          -> "slice"
+  | Clength         -> "length"
   (* list *)
   | Chead           -> "head"
   | Ctail           -> "tail"