+++ src/printer_model.ml
@@ -698,6 +698,20 @@ let pp_mterm fmt (mt : mterm) =
       Format.fprintf fmt "abs (%a)"
         f a
 
+    | Mconcat (x, y) ->
+      Format.fprintf fmt "concat (%a, %a)"
+        f x
+        f y
+
+    | Mslice (x, s, e) ->
+      Format.fprintf fmt "slice (%a, %a, %a)"
+        f x
+        f s
+        f e
+
+    | Mlength x ->
+      Format.fprintf fmt "length (%a)"
+        f x
 
     (* crypto functions *)
 
@@ -718,14 +732,6 @@ let pp_mterm fmt (mt : mterm) =
         f k
         f s
         f x
-
-
-    (* internal functions *)
-
-    | Mstrconcat (l, r)->
-      Format.fprintf fmt "str_concat (%a, %a)"
-        f l
-        f r
 
 
     (* constants *)
@@ -1092,9 +1098,12 @@ let pp_api_list fmt = function
   | Lnth t      -> Format.fprintf fmt "list_nth\t %a" pp_type t
 
 let pp_api_builtin fmt = function
-  | MinBuiltin t -> Format.fprintf fmt "min on %a" pp_type t
-  | MaxBuiltin t -> Format.fprintf fmt "max on %a" pp_type t
-  | AbsBuiltin t -> Format.fprintf fmt "abs on %a" pp_type t
+  | Bmin    t -> Format.fprintf fmt "min on %a"    pp_type t
+  | Bmax    t -> Format.fprintf fmt "max on %a"    pp_type t
+  | Babs    t -> Format.fprintf fmt "abs on %a"    pp_type t
+  | Bconcat t -> Format.fprintf fmt "concat on %a" pp_type t
+  | Bslice  t -> Format.fprintf fmt "slice on %a"  pp_type t
+  | Blength t -> Format.fprintf fmt "length on %a" pp_type t
 
 let pp_api_internal fmt = function
   | RatEq        -> Format.fprintf fmt "rat_eq"
@@ -1102,7 +1111,6 @@ let pp_api_internal fmt = function
   | RatArith     -> Format.fprintf fmt "rat_arith"
   | RatUminus    -> Format.fprintf fmt "rat_uminus"
   | RatTez       -> Format.fprintf fmt "rat_to_tez"
-  | StrConcat    -> Format.fprintf fmt "str_concat"
 
 let pp_api_item_node fmt = function
   | APIAsset      v -> pp_api_asset    fmt v